Restructure DisassemblyManager
